+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Changes compared to v0.7.5.5111
 
 - The ./ prefix in .gitignore has been removed from every single line there.
+- The Icelake-Client x86/x64 CPU, which has been removed from QEMU 7.1, is depreciated on EmuGUI now. Please change the CPU on affected VMs as soon as possible.
 
 # EmuGUI v0.7.5.5111 (based on v0.7.4.5110)